refactor(processor): final design
